+++ GUI/AchievementsFrame.lua
@@ -386,7 +386,9 @@ function achievementsFrame:DisplayAchievement(button, achievement, index, select
 		button.check:Show();
 		button.label:SetWidth(button.label:GetStringWidth() + 4); -- This +4 here is to fudge around any string width issues that arize from resizing a string set to its string width. See bug 144418 for an example.
 		button.tracked:SetChecked(true);
-		button.tracked:Show();
+		if not compact then
+			button.tracked:Show();
+		end
 	else
 		button.check:Hide();
 		button.tracked:SetChecked(false);